+++ CHANGES.rst
@@ -5,6 +5,7 @@ Version 0.9.3 (UNRELEASED)
 --------------------------
 
 - Fixes the verbs used to describe changes to the status of a workflow in order to avoid incorrect phrases such as ``workflow has been failed``.
+- Fixes the loading of Snakemake and CWL workflow specifications when no parameters are specified.
 
 Version 0.9.2 (2023-02-10)
 --------------------------